% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{conan_compare}
\alias{conan_compare}
\title{Compare conan installations}
\usage{
conan_compare(path_lib, curr = 0, prev = -1)
}
\arguments{
\item{path_lib}{Path to the library to compare}

\item{curr}{The previous installation to compare against. Can be a
name (see \link{conan_list} to get names), a negative
number where \code{-n} indicates "\code{n} installations ago" or a
positive number where \code{n} indicates "the \code{n}th
installation". The default value of 0 corresponds to the current
installation.}

\item{prev}{The previous installation to compare against. Can be a
name (see \link{conan_list} to get names), a negative
number where \code{-n} indicates "\code{n} installations ago" or a
positive number where \code{n} indicates "the \code{n}th installation".
The default of -1 indicates the previous installation. Must
refer to an installation before \code{curr}. Use \code{NULL} or -Inf` if
you want to compare against the empty installation.}
}
\value{
An object of class \code{conan_compare}, which can be printed
nicely.
}
\description{
Compare conan installations.
}
