% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{conan_configure}
\alias{conan_configure}
\title{Configuration for conan}
\usage{
conan_configure(
  method,
  ...,
  path_lib,
  path_bootstrap,
  cran = NULL,
  delete_first = FALSE,
  path = ".",
  envvars = NULL
)
}
\arguments{
\item{method}{The method to use; currently \code{script},
\code{pkgdepends}, \code{auto} and \code{renv} are supported.}

\item{...}{Additional arguments, method specific. See Details.}

\item{path_lib}{The library to install into. Could be an absolute
or a relative path.}

\item{path_bootstrap}{The path to a bootstrap library to use. This
needs to contain all the packages required for the method you
are using. For \code{script} this is just \code{remotes}, but for
\code{pkgdepends} it must contain the full recursive dependencies of
\code{pkgdepends}.}

\item{cran}{URL for use as the CRAN repo. If not given we will use
the RStudio CRAN mirror.  This option has no effect when using
renv, as the URLs in your lock file determine the locations that
packages are fetched from.  The intended use of this option is
for where a CRAN repo is misbehaving (e.g., returning 500
errors, or has an invalid/incomplete/out of date index).  The
most likely alternative version to use is \code{cran = "https://cran.r-project.org"}}

\item{delete_first}{Should we delete the library before installing
into it?}

\item{path}{Path to the root where you would run conan from;
typically this is the same path is the root of the project,
often as the working directory.}

\item{envvars}{Environment variables to set before running the
installation.  See Details for format.}
}
\value{
A list with class \code{conan_config}. Do not modify
this object.
}
\description{
Configuration for running conan. Some common options and some
specific to different provisioning methods.
}
\details{
Different methods support different additional arguments:
\itemize{
\item method \code{script} supports the argument \code{script}, which is the
name of the script to run, defaults to "provision.R"
\item method \code{pkgdepends} supports the arguments \code{refs}, which can be
a character vector of references (rather than reading from the
file \code{pkgdepends.txt}) and \code{policy} which is passed through to
\code{\link[pkgdepends:pkg_installation_proposal]{pkgdepends::new_pkg_installation_proposal()}}.
\item method \code{auto} takes an argument \code{environment} which contains a
list of packages to install and source files to scan for
dependencies.
\item method \code{renv} takes no arguments.
}

Setting environment variables while running the installation comes
uses the \code{envvars} argument; this system is designed to play well
with \code{hipercow}, though it does not require it.  We expect a
\code{data.frame} with columns \code{name}, \code{value} and (optionally)
\code{secret}.  If \code{secret} is given, it must be a logical value
indicating that \code{value} has been encrypted with an \code{rsa} public
key.  If any \code{secret} is \code{TRUE}, then \code{envvvars} must also have an
\emph{attribute} \code{key} that contains the path to private rsa key to
decrypt the secrets (i.e., \code{attr(envvars, "key")}).  If you use
secret environment variables, then the \code{openssl} package must be
present in conan's bootstrap.
}
