% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{conan_describe}
\alias{conan_describe}
\title{Describe a library}
\usage{
conan_describe(path_lib)
}
\arguments{
\item{path_lib}{Path to the library}
}
\value{
A list with class \code{conan_describe}. We'll write some
tooling to work with these soon!
}
\description{
Describe a library.  This creates a summary of version information
from a library.  This may be slightly slow on network filesystems
with large libraries.
}
