% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{conan_is_current}
\alias{conan_is_current}
\title{Test if a conan installation is current}
\usage{
conan_is_current(path_lib, config)
}
\arguments{
\item{path_lib}{Path to the library to compare}

\item{config}{A configuration object}
}
\value{
A boolean, \code{TRUE} if we are up to date and \code{FALSE} otherwise
}
\description{
Test if a conan installation is current. This just checks if the
instructions to provision have changed since we last ran, not if
the installation \emph{would} do anything different.
}
