% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{conan_list}
\alias{conan_list}
\title{Test if a conan installation is current}
\usage{
conan_list(path_lib)
}
\arguments{
\item{path_lib}{Path to the library to compare}
}
\value{
A \link{data.frame} with columns:
\itemize{
\item \code{name}: the name of the installation. This might be useful with
\code{conan_compare}
\item \code{time}: the time the installation was started
\item \code{hash}: the installation hash
\item \code{method}: the method used for the installation
\item \code{args}: the arguments to the installation (as a list column)
}

This object also has class \code{conan_list} so that it prints nicely,
but you can drop this with \code{as.data.frame}.
}
\description{
List conan installations, and optionally test if they are current.
}
