% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{conan_list_installations}
\alias{conan_list_installations}
\title{List conan installations}
\usage{
conan_list_installations(path_lib)
}
\arguments{
\item{path_lib}{The library that conan has installed into}
}
\value{
Invisibly, a character vector of names
}
\description{
List installations that have been carried out by conan into a
library
}
